park and ride test scenario
git-svn-id: file:///srv/local/data/cleanRepo/trunk@20032 a0517364-e873-4167-b1c8-416daa5a6b91

+++ playgrounds/ikaddoura/src/main/java/playground/ikaddoura/parkAndRide/pRstrategy/ParkAndRidePlanMutateTimeAllocation.java
@@ -42,11 +42,11 @@ import playground.ikaddoura.parkAndRide.pR.ParkAndRideConstants;
  */
 public class ParkAndRidePlanMutateTimeAllocation implements PlanAlgorithm {
 
-	private final int mutationRange;
+	private final Double mutationRange;
 	private final Random random;
 	private boolean useActivityDurations;
 
-	public ParkAndRidePlanMutateTimeAllocation(final int mutationRange, final Random random) {
+	public ParkAndRidePlanMutateTimeAllocation(final Double mutationRange, final Random random) {
 		this.mutationRange = mutationRange;
 		this.random = random;
 	}
